requirements.txt
Pillow=9.2.0
boto3==1.24.86